Use GutenbergEditorComponent()
