+++ lib/gutenberg/gutenberg-editor-component.js
@@ -229,4 +229,41 @@ export default class GutenbergEditorComponent extends AsyncBaseContainer {
 		);
 		return hasPublishButton && ! hasRevertButton;
 	}
+
+	async schedulePost( publishDate ) {
+		await driverHelper.clickWhenClickable(
+			this.driver,
+			By.css( '.editor-post-publish-panel__toggle' )
+		);
+		await driverHelper.waitTillPresentAndDisplayed(
+			this.driver,
+			By.css( '.editor-post-publish-panel__header' )
+		);
+		await driverHelper.verifyTextPresent(
+			this.driver,
+			By.css( '.editor-post-publish-panel__link' ),
+			publishDate
+		);
+		await driverHelper.clickWhenClickable(
+			this.driver,
+			By.css( '.editor-post-publish-panel__header-publish-button button:not([disabled])' )
+		);
+		await driverHelper.waitTillNotPresent(
+			this.driver,
+			By.css( '.editor-post-publish-panel__content .components-spinner' )
+		);
+		await driverHelper.waitTillPresentAndDisplayed(
+			this.driver,
+			By.css( '.editor-post-publish-panel__header-published' )
+		);
+		return await driverHelper.verifyTextPresent(
+			this.driver,
+			By.css( '.editor-post-publish-panel__header-published' ),
+			'Scheduled'
+		);
+	}
+
+	async closeScheduledPanel() {
+		await driverHelper.clickWhenClickable( this.driver, By.css( '.dashicons-no-alt' ) );
+	}
 }
